Rename package to azureventhubs
