+++ clients/pkg/promtail/targets/azureeventhubs/parser.go
@@ -1,4 +1,4 @@
-package azureeventhub
+package azureeventhubs
 
 import (
 	"bytes"
