+++ clients/pkg/promtail/targets/azureeventhubs/parser_test.go
@@ -1,4 +1,4 @@
-package azureeventhub
+package azureeventhubs
 
 import (
 	"os"
