+++ clients/pkg/promtail/targets/azureeventhubs/target_syncer_test.go
@@ -1,4 +1,4 @@
-package azureeventhub
+package azureeventhubs
 
 import (
 	"errors"
