+++ clients/pkg/promtail/targets/azureeventhubs/targetmanager.go
@@ -1,4 +1,4 @@
-package azureeventhub
+package azureeventhubs
 
 import (
 	"github.com/go-kit/log"
