+++ clients/pkg/promtail/targets/azureeventhubs/testdata/custom_message_and_logic_app_logs.json
