+++ clients/pkg/promtail/targets/azureeventhubs/testdata/custom_payload_json_with_records_string.json
