+++ clients/pkg/promtail/targets/azureeventhubs/testdata/custom_payload_text.txt
