+++ clients/pkg/promtail/targets/azureeventhubs/testdata/function_app_logs_message.txt
