+++ clients/pkg/promtail/targets/azureeventhubs/testdata/logic_app_logs_message.json
