+++ clients/pkg/promtail/targets/azureeventhubs/testdata/message_with_invalid_time.json
